replaced pid_t with unsigned int for cross compatibility

+++ include/ctm/ctm.h
@@ -29,7 +29,7 @@ unsigned long memory_get_total(const Memory* memory);
 unsigned long memory_get_free(const Memory* memory);
 unsigned long memory_get_available(const Memory* memory);
 
-Process*     process_get(unsigned long pid);
+Process*     process_get(unsigned int pid);
 void         process_free(Process* process);
 unsigned int process_get_pid(const Process* process);
 const char*  process_get_name(const Process* process);
